refactor: Move constants out of `colorizer/types.ts` and into a separate file
