+++ src/colorizer/utils/data_utils.ts
@@ -1,6 +1,7 @@
 import { MAX_FEATURE_CATEGORIES } from "../../constants";
+import { isThresholdCategorical, isThresholdNumeric } from "../../constants";
 import { ColorRampData } from "../colors/color_ramps";
-import { FeatureThreshold, isThresholdCategorical, isThresholdNumeric, ThresholdType } from "../types";
+import { FeatureThreshold, ThresholdType } from "../types";
 
 import ColorRamp from "../ColorRamp";
 import Dataset, { FeatureType } from "../Dataset";
