+++ src/components/Export.tsx
@@ -17,7 +17,8 @@ import styled from "styled-components";
 import { clamp } from "three/src/math/MathUtils";
 
 import { ExportIconSVG } from "../assets";
-import { defaultViewerConfig, ViewerConfig } from "../colorizer/types";
+import { defaultViewerConfig } from "../colorizer/constants";
+import { ViewerConfig } from "../colorizer/types";
 import { AnalyticsEvent, triggerAnalyticsEvent } from "../colorizer/utils/analytics";
 import { FlexColumn, FlexColumnAlignCenter, FlexRow } from "../styles/utils";
