+++ src/components/Tabs/FeatureThresholdsTab.tsx
@@ -9,13 +9,12 @@ import { Dataset } from "../../colorizer";
 import {
   CategoricalFeatureThreshold,
   FeatureThreshold,
-  isThresholdCategorical,
-  isThresholdNumeric,
   NumericFeatureThreshold,
   ThresholdType,
 } from "../../colorizer/types";
 import { thresholdMatchFinder } from "../../colorizer/utils/data_utils";
 import { ScrollShadowContainer, useScrollShadow } from "../../colorizer/utils/react_utils";
+import { isThresholdCategorical, isThresholdNumeric } from "../../constants";
 import { MAX_FEATURE_CATEGORIES } from "../../constants";
 import { FlexColumn } from "../../styles/utils";
