+++ src/components/Tabs/Settings/VectorFieldSettings.tsx
@@ -3,7 +3,8 @@ import { Card, Checkbox, ColorPicker } from "antd";
 import React, { ReactElement } from "react";
 import { Color, ColorRepresentation } from "three";
 
-import { VECTOR_KEY_MOTION_DELTA, VectorConfig, ViewerConfig } from "../../../colorizer/types";
+import { VECTOR_KEY_MOTION_DELTA } from "../../../colorizer/constants";
+import { VectorConfig, ViewerConfig } from "../../../colorizer/types";
 import { DEFAULT_OUTLINE_COLOR_PRESETS } from "./constants";
 
 import SelectionDropdown from "../../Dropdowns/SelectionDropdown";
